Powershell AD U groups and Mirror Permissions
In this commit I have added two scripts that are great automation tools. One allows you to find a U Group by keyword no matter the full U group name if it has your keyword the console will output it to you.

The second script is a method to mirror permissions on a Computer Device