Foundation Scripts (Learning)
So I'm learning PowerShell and this is a little project I've started to try and understand the platform better.

This ps1 allows you to start a program in which text is generated and specific data is gathered and then saved to a file. There are some checks that have been done just in case specific paths cannot be found or seeing if the name of a file is already in paths.

There are some pre excisting bugs like the file name not being detected correctly outside of vscode but this is something I can return to and fix later