AD DC Lab Creation
I have made a AD DC Lab - This is a Virtual machine on VMware that is using the Win 2022 Server ISO image. I have set up Active Directory as with this I plan to make multiple projects. One being related to the Tech Buddy Script to save made files onto the server in a shared drive.